returns appropriate responses

+++ src/main/java/com/demo/student_management/service/StudentService.java
@@ -3,6 +3,8 @@ package com.demo.student_management.service;
 import com.demo.student_management.model.Student;
 import com.demo.student_management.repository.StudentRepository;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -13,19 +15,41 @@ public class StudentService {
     @Autowired
     private StudentRepository studentRepository;
 
-    public List<Student> getAllStudents() {
-        return studentRepository.findAll();
+    public ResponseEntity<String> getAllStudents() {
+        List<Student> l = studentRepository.findAll();
+        String mssg;
+
+        if (l.size() < 1)
+            mssg = "No student registered";
+        else
+            mssg = l.toString();
+
+        return new ResponseEntity<>(mssg, HttpStatus.OK);
     }
 
-    public Student saveStudent(Student student) {
-        return studentRepository.save(student);
+    public ResponseEntity<String> saveStudent(Student student) {
+        studentRepository.save(student);
+        return new ResponseEntity<>("Student Registered", HttpStatus.OK);
     }
 
-    public Student getStudentById(Long id) {
-        return studentRepository.findById(id).orElse(null);
+    public ResponseEntity<String> getStudentById(Long id) {
+        Student s = studentRepository.findById(id).orElse(null);
+
+        if (s == null)
+            return new ResponseEntity<>("No such student ID exists",
+                    HttpStatus.BAD_REQUEST);
+
+        return new ResponseEntity<>(s.toString(), HttpStatus.OK);
     }
 
-    public void deleteStudentById(Long id) {
+    public ResponseEntity<String> deleteStudentById(Long id) {
+        Student s = studentRepository.findById(id).orElse(null);
+
+        if (s == null)
+            return new ResponseEntity<>("No such student ID exists",
+                    HttpStatus.BAD_REQUEST);
+
         studentRepository.deleteById(id);
+        return new ResponseEntity<>("Student deleted", HttpStatus.OK);
     }
 }
